backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiomysql
requests
apscheduler
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-multipart
httpx
